test(do): split write-throughput bench into isolated one-scenario files
CodSpeed's instrumented runner re-runs each bench body many times and shares the
suite's fixtures across benches (beforeAll once, beforeEach per-run) — so the
patch/replace benches lost their seed row to sibling churn on the shared writer
("document not found: seed"), and no in-body guard fixed it reliably.

Adopt the visulima `__bench__` convention: one scenario per file, each owning its
own isolated writer, with the common writer factory in `shared.ts`. The five
write-throughput scenarios (bare/agg/rank insert, patch, replace) become five
files; patch/replace seed their row once in beforeAll and never delete it, so it
survives the re-runs.
